updated execution status and update next models if exists

+++ cmmmodel/saveJobResponse.py
@@ -15,6 +15,7 @@ def saveModelResponse(external_id, std_out, std_err):
 
     execution_obj = ModelExecutionHistory.objects.get(externalId=external_id)
     execution_obj.end=timezone.now()
+    execution_obj.status=ModelExecutionHistory.OK
     execution_obj.answer=std_out
     execution_obj.error += std_err
     execution_obj.save()
@@ -23,8 +24,9 @@ def saveModelResponse(external_id, std_out, std_err):
     next_models = ModelExecutionQueue.objects.filter(modelExecutionHistory=execution_obj).order_by('id').\
         values_list('id', flat=True)
 
-    Run().runModel(None, execution_obj.scene_id, next_models[0], next_models[1:])
-    ModelExecutionQueue.objects.filter(modelExecutionHistory=execution_obj).delete()
+    if len(next_models) > 1:
+        Run().runModel(None, execution_obj.scene_id, next_models[0], next_models[1:])
+        ModelExecutionQueue.objects.filter(modelExecutionHistory=execution_obj).delete()
 
 if __name__ == "__main__":
     """ update execution record """
